+++ website/content/content/index.ts
@@ -6,6 +6,9 @@ import { colorsLight } from '../../temp/defaultValues'
 import * as avatarContent from './avatar';
 import * as badgeContent from './badge';
 import * as cardContent from './card';
+import * as imageContent from './image';
+import * as carouselContent from './carrousel';
+import * as collapseContent from './collapse';
 
 export default {
   iconContent,
@@ -15,5 +18,8 @@ export default {
   avatarContent,
   badgeContent,
   cardContent,
+  imageContent,
+  carouselContent,
+  collapseContent,
   themeColors: colorsLight,
 };