+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.0-beta.1.4.8
+
+`2022-12-01`
+
+- Added Carousel documentation page.
+- Added Collapse documentation page.
+- Added Image documentation page.
+- Changed width context from windows to screen.
+- Fixed openKeys prop in collapse.
+
 ## 1.0.0-beta.1.4.7
 
 `2022-12-01`